requirement.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.3.0
auth0-python==4.7.2
Authlib==1.3.1
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.8
cohere==5.13.5
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.0
contourpy==1.2.0
crewai==0.95.0
crewai-tools==0.25.8
cryptography==43.0.3
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.15
deprecation==2.1.0
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
docx2txt==0.8
durationpy==0.9
embedchain==0.1.126
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.115.6
fastavro==1.10.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.47.2
frozenlist==1.5.0
fsspec==2024.12.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.0
google-auth==2.37.0
google-cloud==0.34.0
google-cloud-aiplatform==1.75.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-language==2.16.0
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
gptcache==0.1.44
greenlet==3.1.1
grpc-google-iam-v1==0.14.0
grpcio==1.69.0
grpcio-health-checking==1.69.0
grpcio-status==1.69.0
grpcio-tools==1.69.0
h11==0.14.0
h2==4.1.0
h5py==3.11.0
hpack==4.0.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.27.1
humanfriendly==10.0
hyperframe==6.0.1
idna==2.8
ijson==3.3.0
imbalanced-learn==0.12.3
imblearn==0.0
importlib_metadata==8.5.0
importlib_resources==6.5.2
iniconfig==2.0.0
instructor==1.7.2
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
json_repair==0.35.0
jsonpatch==1.33
jsonpickle==4.0.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.0
jupyter_core==5.5.0
kaggle==1.6.17
keras==3.5.0
kiwisolver==1.4.5
kubernetes==31.0.0
lancedb==0.17.0
langchain==0.3.15
langchain-cohere==0.3.4
langchain-community==0.3.14
langchain-core==0.3.31
langchain-experimental==0.3.4
langchain-openai==0.2.14
langchain-text-splitters==0.3.4
langsmith==0.1.147
libclang==18.1.1
linkup-sdk==0.2.1
litellm==1.57.0
llvmlite==0.42.0
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mem0ai==0.1.40
ml-dtypes==0.4.1
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.21.1
nest-asyncio==1.5.8
networkx==3.4.2
nodeenv==1.9.1
npm==0.1.1
numba==0.59.0
numpy==1.26.3
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.59.3
openpyxl==3.1.5
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
opt-einsum==3.3.0
optional-django==0.1.0
optree==0.12.1
orjson==3.10.13
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.0
parameterized==0.9.0
parso==0.8.3
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==10.2.0
platformdirs==4.1.0
pluggy==1.5.0
portalocker==2.10.1
posthog==3.7.5
# Editable install with no version control (project_aiver2==0.1.0)
-e c:\ai\project_aiver2
prompt-toolkit==3.0.43
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
psutil==5.9.6
pure-eval==0.2.2
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
pygame==2.5.1
Pygments==2.17.2
PyJWT==2.10.1
pylance==0.20.0
pyparsing==3.1.1
pypdf==5.1.0
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pyright==1.1.391
pysbd==0.3.4
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.8.2
python-dotenv==1.0.1
python-slugify==8.0.4
pytube==15.0.0
pytz==2024.1
pyvis==0.3.2
pywin32==306
PyYAML==6.0.2
pyzmq==25.1.2
qdrant-client==1.12.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.8.1
rpds-py==0.22.3
rsa==4.9
schema==0.7.7
scikit-learn==1.5.2
scipy==1.12.0
scrapegraph_py==1.8.0
seaborn==0.13.2
selenium==4.27.1
serpapi==0.1.5
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sns==0.1
sortedcontainers==2.4.0
soupsieve==2.6
sparse==0.15.1
spider-client==0.1.25
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.41.3
streamlit==1.41.1
sympy==1.12
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.20.3
toml==0.10.2
tomli==2.2.1
tomli_w==1.1.0
tornado==6.4
tqdm==4.67.1
traitlets==5.14.0
trio==0.28.0
trio-websocket==0.11.1
typer==0.15.1
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.4
ucimlrepo==0.0.7
urllib3==2.3.0
uv==0.5.14
uvicorn==0.34.0
validators==0.34.0
watchdog==6.0.0
watchfiles==1.0.3
wcwidth==0.2.12
weaviate-client==4.10.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.0.4
wrapt==1.16.0
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
